requirements.txt
Django==1.11.1
psycopg2==2.7.1
unholster.milieu==1.0.0
django-password-reset==0.9
redis==2.10.5
channels==1.1.3
asgi-redis==1.3.0
requests==2.14.2
celery==4.0.2
django-celery-beat==1.0.1
raven==6.0.0
django-rest-swagger==2.1.2
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
